feat(#56) : Reservaton 타입 분리
